Updated file structure and updated gitignore
