+++ .gitignore
@@ -1,5 +1,3 @@
-api/node_modules
-client/node_modules
-rtmpserver/node_modules
+node_modules
 .vscode/
-client/build
+build
