+++ src/assets/thumbnails.json
