+++ src/components/Carousel.js
