+++ src/components/streams/StreamDelete.js
