+++ src/reducers/streamReducer.js
