+++ src/reducers/streamServerReducer.js
